Improve static analysis of stack accesses
Memoro is a heap profiler and is not interested in the stack.
Not instrumenting stack accesses is really important to avoid wasting
run-time CPU to discard those samples.

The previous approach was using `GetPointerOperand()` to go one step
back in the SSA graph, in order to determine if the access is targeting
the stack.

The new approach uses `GetUnderlyingObject()` to go multiple steps back
in the SSA and can find more opportunities to discard.
It also check for global variables, which we are also not interested in.